firmata v0.0.2 Firmata.Board

This module describes the board that is being communicated with and is the primary module that a user will interact with.

Pin Modes

Mode

@input

@output

@analog

@pwm

@servo

@shift

@i2c

@onewire

@stepper

@serial

@ignore

@ping_read

@sonar

@unknown

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

digital_write(board, pin, value)

init(arg)

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

init_arg is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state} except handle_info(:timeout, state) will be called after timeout milliseconds if no messages are received within the timeout.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

neopixel(board, index, arg)

neopixel_brightness(board, brightness)

neopixel_register(board, pin, num_pixels)

pin_state(board, pin)

report_analog_channel(board, channel, value)

Requests that the board 'report' the value on channel to this process. It will send messages of the format: {:firmata, {:analog_read, channel, value}}

set_pin_mode(board, pin, mode)

Sets the pin to the given mode.

sonar_config(board, trigger, echo, max_distance, ping_interval)

start_link(port, opts \\ [], name \\ nil)

Connects to the firmata host at port.

stop(pid)

sysex_write(board, cmd, data)

to_hex(arg)

Converts a byte to a string hex representation.

Returns the hex representation of byte

Examples:

iex> Firmata.Protocol.Mixin.to_hex(<<10>>) "0xA"